socket modify
