+++ socketrun.py
@@ -1,0 +1,46 @@
+from flask import render_template, request, session
+
+from planet import Flask, socketio
+from planet.route.RouteSocket import Mynamespace
+
+app = Flask(__name__)
+
+
+@app.route('/api/v2/test')
+def hi():
+    #     return 'ok'
+    return render_template('index.html')
+
+
+@app.route('/api/v2/mes')
+def mes():
+    #     return 'ok'
+    event_name = 'test'
+    data = request.args.get("msg")
+    roomid = request.args.get('room')
+    broadcasted_data = {'data': data}
+    print("publish msg==>", broadcasted_data)
+    # socketio.emit(event_name, broadcasted_data, broadcast=True)
+    socketio.emit(event_name, broadcasted_data, room=roomid)
+    return 'send msg successful!'
+
+
+@socketio.on('json')
+def handle_json(json):
+    print(session.get('id'))
+    # session['id'] = 'json'
+    print('received json: ' + str(json))
+
+
+@socketio.on('message')  # 接收匿名send消息
+def message_handler(*args):
+    print(session.get('id'))
+    # session['id'] = 'message'
+    print('recive {0} data from client'.format(type(args[0])), args)
+    return args
+
+
+if __name__ == '__main__':
+    socketio.init_app(app)
+    socketio.on_namespace(Mynamespace('/'))
+    socketio.run(app, port=7444)
